REF/TST: collect test by methods
